api_server: Bind a session to its nonce
Generate the nonce when creating the session.
